backend/requirements.txt
Django==6.0.2
python-dotenv==1.2.1
retell-sdk==5.16.0